chore(tauri-api): move constants
