+++ apps/native/src/components/widget/widget.test.tsx
@@ -17,8 +17,6 @@ vi.mock("@/tauri-api", () => ({
   ipcRenderer: {
     on: vi.fn().mockReturnValue(Promise.resolve(() => {})),
   },
-  CONFIG_CHANGED_CHANNEL: "config-changed",
-  DEFAULT_MAX_ITERATIONS: 25,
 }));
 
 // Mock hooks
