+++ apps/native/src/hooks/use-evolve.ts
@@ -1,5 +1,6 @@
 import { useWidgetStore } from "@/stores/widget-store";
-import { darwinAPI, EVOLVE_EVENT_CHANNEL, ipcRenderer } from "@/tauri-api";
+import { EVOLVE_EVENT_CHANNEL } from "@/../shared/constants";
+import { darwinAPI, ipcRenderer } from "@/tauri-api";
 import type { EvolveEvent } from "@/types/shared";
 import { useCallback } from "react";
 import { useGitOperations } from "./use-git-operations";
